Added (unsigned int) cast to get rid of gcc-2.6.0 warning: Linker.cc:374: warning: unsigned int format, pointer arg (arg 7)

+++ trunk/ptclassic/src/kernel/Linker.cc
@@ -371,7 +371,8 @@ int Linker::multiLink (int argc, char** argv) {
 	char command[512];
 	// DYNLIB is hppa specific
 	sprintf (command, "%s %s -A %s %s %x %s -o ",
-		 LOADER, LOADOPTS, symTableName, LOC_OPT, availMem, DYNLIB);
+		 LOADER, LOADOPTS, symTableName, LOC_OPT,
+		 (unsigned int)availMem, DYNLIB);
 	D( Error::message(command); )
 	StringList cmd = command;
 	cmd << tname;
